# docs/swagger.yaml
basePath: /api
consumes:
- application/json
definitions:
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  models.CreateNotificationRequest:
    properties:
      action_url:
        type: string
      customer_id:
        type: integer
      icon_name:
        type: string
      is_read:
        type: boolean
      message:
        type: string
      order_id:
        type: integer
      payment_id:
        type: integer
      priority:
        type: string
      read_at:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
...
  models.NotificationResponse:
    properties:
      action_url:
        type: string
      created_at:
        type: string
      customer_id:
        type: integer
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      icon_name:
        type: string
      id:
        type: integer
      is_read:
        type: boolean
      message:
        type: string
      order_id:
        type: integer
      payment_id:
        type: integer
      priority:
        type: string
      read_at:
        $ref: '#/definitions/types.DateTime'
      title:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  models.NotificationSelectOption:
    properties:
      id:
        type: integer
      name:
        description: From Title field
        type: string
    type: object
...
  models.UpdateNotificationRequest:
    properties:
      action_url:
        type: string
      customer_id:
        type: integer
      icon_name:
        type: string
      is_read:
        type: boolean
      message:
        type: string
      order_id:
        type: integer
      payment_id:
        type: integer
      priority:
        type: string
      read_at:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
...
  types.DateTime:
    properties:
      time.Time:
        type: string
    type: object
  types.ErrorResponse:
    properties:
      details: {}
      error:
        type: string
      success:
        type: boolean
    type: object
  types.PaginatedResponse:
    properties:
      data: {}
      pagination:
        $ref: '#/definitions/types.Pagination'
    type: object
  types.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  types.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact:
    email: info@negenet.com
    name: Negenet Team
    url: https://negenet.com
  description: This is the API documentation for Negenet
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://negenet.com/terms
  title: Negenet API
  version: 2.0.0
paths:
...
  /notifications:
    get:
      consumes:
      - application/json
      description: Get a list of notifications
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Sort field (id, created_at, updated_at,title,message,type,priority,is_read,read_at,customer_id,order_id,payment_id,action_url,icon_name,)
        in: query
        name: sort
        type: string
      - description: Sort order (asc, desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.PaginatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: List notifications
      tags:
      - App/Notification
    post:
      consumes:
      - application/json
      description: Create a new Notification with the input payload
      parameters:
      - description: Create Notification request
        in: body
        name: notifications
        required: true
        schema:
          $ref: '#/definitions/models.CreateNotificationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.NotificationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Create a new Notification
      tags:
      - App/Notification
  /notifications/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Notification by its id
      parameters:
      - description: Notification id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Delete a Notification
      tags:
      - App/Notification
    get:
      consumes:
      - application/json
      description: Get a Notification by its id
      parameters:
      - description: Notification id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.NotificationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get a Notification
      tags:
      - App/Notification
    put:
      consumes:
      - application/json
      description: Update a Notification by its id
      parameters:
      - description: Notification id
        in: path
        name: id
        required: true
        type: integer
      - description: Update Notification request
        in: body
        name: notifications
        required: true
        schema:
          $ref: '#/definitions/models.UpdateNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.NotificationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Update a Notification
      tags:
      - App/Notification
  /notifications/all:
    get:
      consumes:
      - application/json
      description: Get a simplified list of all notifications with id and name only
        (for dropdowns/select boxes)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.NotificationSelectOption'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: List all notifications for select options
      tags:
      - App/Notification
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: API Key for authentication
    in: header
    name: X-Api-Key
    type: apiKey
  BearerAuth:
    description: Enter your token with the prefix "Bearer "
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"